// codegen/crm/properties/model/batchInputPropertyCreate.ts
/**
 * Properties
 * All HubSpot objects store data in default and custom properties. These endpoints provide access to read and modify object properties in HubSpot.
 *
 * The version of the OpenAPI document: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { PropertyCreate } from './propertyCreate';

export class BatchInputPropertyCreate {
    'inputs': Array<PropertyCreate>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "inputs",
            "baseName": "inputs",
            "type": "Array<PropertyCreate>"
        }    ];

    static getAttributeTypeMap() {
        return BatchInputPropertyCreate.attributeTypeMap;
    }
}