// codegen/crm/properties/model/batchInputPropertyName.ts
/**
 * Properties
 * All HubSpot objects store data in default and custom properties. These endpoints provide access to read and modify object properties in HubSpot.
 *
 * The version of the OpenAPI document: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { PropertyName } from './propertyName';

export class BatchInputPropertyName {
    'inputs': Array<PropertyName>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "inputs",
            "baseName": "inputs",
            "type": "Array<PropertyName>"
        }    ];

    static getAttributeTypeMap() {
        return BatchInputPropertyName.attributeTypeMap;
    }
}