// codegen/crm/properties/model/optionInput.ts
/**
 * Properties
 * All HubSpot objects store data in default and custom properties. These endpoints provide access to read and modify object properties in HubSpot.
 *
 * The version of the OpenAPI document: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class OptionInput {
    /**
    * A human-readable option label that will be shown in HubSpot.
    */
    'label': string;
    /**
    * The internal value of the option, which must be used when setting the property value through the API.
    */
    'value': string;
    /**
    * A description of the option.
    */
    'description'?: string;
    /**
    * Options are shown in order starting with the lowest positive integer value. Values of -1 will cause the option to be displayed after any positive values.
    */
    'displayOrder'?: number;
    /**
    * Hidden options won\'t be shown in HubSpot.
    */
    'hidden': boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "label",
            "baseName": "label",
            "type": "string"
        },
        {
            "name": "value",
            "baseName": "value",
            "type": "string"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "displayOrder",
            "baseName": "displayOrder",
            "type": "number"
        },
        {
            "name": "hidden",
            "baseName": "hidden",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return OptionInput.attributeTypeMap;
    }
}